requirements.txt
Flask==3.0.0
flask-cors==4.0.0
fpdf2==2.7.7
Werkzeug==3.0.1
